+++ src/dfwimpl/state_driver.cpp
@@ -59,14 +59,32 @@ void state_driver::init(dfw::kernel&) {}
 
 void state_driver::prepare_video(dfw::kernel& kernel) {
 
+	const auto& argman=kernel.get_arg_manager();
+
+	int w=config.int_from_path("video:window_w_px");
+	int h=config.int_from_path("video:window_h_px");
+
+	if(argman.exists("-w") && argman.arg_follows("-w")) {
+
+		const std::string& window_size_str=argman.get_following("-w");
+
+		auto xpos=window_size_str.find("x");
+		if(std::string::npos==xpos) {
+			throw std::runtime_error("-w parameter must be specified in wxh");
+		}
+
+		w=std::stoi(window_size_str.substr(0, xpos));
+		h=std::stoi(window_size_str.substr(xpos+1));
+
+		lm::log(log).info()<<"window size specified by command line as "<<window_size_str<<" ["<<w<<"x"<<h<<"]"<<std::endl;
+	}
+
 	std::stringstream ss;
 	ss<<config.string_from_path("video:window_title")<<" v"<<MAJOR_VERSION<<"."<<MINOR_VERSION<<"."<<PATCH_VERSION<<"-"<<BUILD_VERSION;
 
 	kernel.init_video_system({
-		config.int_from_path("video:window_w_px"),
-		config.int_from_path("video:window_h_px"),
-		config.int_from_path("video:window_w_logical"),
-		config.int_from_path("video:window_h_logical"),
+		w,h,
+		w,h,
 		ss.str(),
 		config.bool_from_path("video:window_show_cursor"),
 		config.get_screen_vsync()
@@ -150,15 +168,15 @@ void state_driver::prepare_resources(dfw::kernel& /*kernel*/) {
 */
 }
 
-void state_driver::register_controllers(dfw::kernel& /*kernel*/) {
+void state_driver::register_controllers(dfw::kernel& _kernel) {
 
 	auto reg=[this](ptr_controller& _ptr, int _i, dfw::controller_interface * _ci) {
 		_ptr.reset(_ci);
 		register_controller(_i, *_ptr);
 	};
 
-	unsigned int    screen_w=config.int_from_path("video:window_w_logical"),
-					screen_h=config.int_from_path("video:window_h_logical");
+	unsigned int    screen_w=_kernel.get_screen().get_w(),
+					screen_h=_kernel.get_screen().get_h();
 
 	reg(
 		c_file_browser,